Fixes a runtime crash on Android 11 (#111)
* Fixes a runtime crash on Android 11

Android 11 seems to have killed the ability to use a `LIMIT` clause, introduced in `VERSION_CODES.O`, in the order argument of a `ContentProvider` query.

This PR changes the query to use a `Bundle` when the version code is >= `VERSION_CODES.O`. It also changes the sorting order to use `DATE_TAKEN` if available, or `DATE_MODIFIED` otherwise.

Target and Build versions have been set to 30.