Fixes android.useAndroidX not being visible to isolated module builds
When building in isolation, the root project's gradle.properties file is not being read. This commit adds the android.useAndroidX property with a true value to the two places where we're doing isolated module builds. This approach was taken because it impacted one file, rather than the alternative of adding module level gradle.properties to the affected modules.